Moved interface and import to correct files.
