+++ ts/Core/Globals.ts
@@ -82,11 +82,6 @@ declare global {
     interface TouchList {
         changedTouches: Array<Touch>;
     }
-    interface Window {
-        canvg: CanvgNamespace;
-        jsPDF: typeof jsPDF;
-        svg2pdf: Function;
-    }
     /**
      * @private
      * @deprecated
@@ -94,7 +89,6 @@ declare global {
      */
     const win: Window|undefined;
 }
-
 
 /* *
  *
