+++ ts/masters/modules/exporting.src.ts
@@ -14,11 +14,9 @@ import Highcharts from '../../Core/Globals.js';
 import '../../Extensions/FullScreen.js';
 import Exporting from '../../Extensions/Exporting/Exporting.js';
 import HttpUtilities from '../../Core/HttpUtilities.js';
-import OfflineExporting from '../../Extensions/OfflineExporting/OfflineExporting.js';
 const G: AnyRecord = Highcharts;
 G.HttpUtilities = HttpUtilities;
 G.ajax = HttpUtilities.ajax;
 G.getJSON = HttpUtilities.getJSON;
 G.post = HttpUtilities.post;
 Exporting.compose(G.Chart, G.Renderer);
-OfflineExporting.compose(G.Chart);
